% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_eurocordex_units.R
\name{convert_eurocordex_units}
\alias{convert_eurocordex_units}
\title{Convert units of columns of a data.frame with the udunits2 package}
\usage{
convert_eurocordex_units(data, station_names, unit_in, unit_out)
}
\arguments{
\item{data}{data.frame: contains a column with data; columns indicated by 'station_names' are converted from 'unit_in' to 'unit_out'}

\item{station_names}{character-array: contains names of the stations (columns of data), which units should be converted}

\item{unit_in}{character: unit of the input data (FROM which should be converted)}

\item{unit_out}{character: unit of the output data (TO which should be converted)}
}
\value{
data.from: same format as the
}
\description{
All columns of 'data', which are listed in 'station_names', are converted from 'unit_in' to 'unit_out' with the help of the udunits2 package.
}
\examples{

# see example in README


}
