% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_varname_eurocordex2swat.R
\name{convert_varname_eurocordex2swat}
\alias{convert_varname_eurocordex2swat}
\title{Convert EURO-CORDEX variable name to SWAT2012 variable name}
\usage{
convert_varname_eurocordex2swat(varname, stop_on_error = TRUE)
}
\arguments{
\item{varname}{character or character-array: one or more EURO-CORDEX variable names}

\item{stop_on_error}{boolean: stop, when one or more entries of varname are not convertible}
}
\value{
character or character-array: SWAT2012 variable names
}
\description{
Convert EURO-CORDEX variable name to SWAT2012 variable name
}
\examples{

 var <- 'hurs'
 var_swat <- convert_varname_eurocordex2swat(var)
 print(var_swat)
 # [1] "hmd"
 
 var <- 'rsds'
 var_swat <- convert_varname_eurocordex2swat(var)
 print(var_swat)
 # [1] "slr"
 
 var <- 'tasmax'
 var_swat <- convert_varname_eurocordex2swat(var)
 print(var_swat)
 # [1] "tmp"
 
 
}
