% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_varname_swat2eurocordex.R
\name{convert_varname_swat2eurocordex}
\alias{convert_varname_swat2eurocordex}
\title{Convert SWAT2012 variable name to EURO-CORDEX variable name}
\usage{
convert_varname_swat2eurocordex(varname, stop_on_error = TRUE)
}
\arguments{
\item{varname}{character or character-array: one or more SWAT2012 variable names}

\item{stop_on_error}{boolean: stop, when one or more entries of varname are not convertible}
}
\value{
character or character-array: EURO-CORDEX variable names
}
\description{
Convert SWAT2012 variable name to EURO-CORDEX variable name
}
\details{
Converts SWAT2012 variable name to EURO-CORDEX variable name.

NOTE: It might return one or MORE values per input value. Please have a look at the examples for details.
}
\examples{

 # PLEASE NOT THE LAST EXAMPLE!

 var <- 'pcp'
 var_ec <- convert_varname_swat2eurocordex(var)
 print(var_ec)
 # [1] "pr"
 
 var <- 'slr'
 var_ec <- convert_varname_swat2eurocordex(var)
 print(var_ec)
 # [1] "rsds"
 
 var <- 'tmp'
 var_ec <- convert_varname_swat2eurocordex(var)
 print(var_ec)
 # [1] "tasmin" "tasmax"
 
 
}
