% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_eurocordex_filelist.R
\name{filter_eurocordex_filelist}
\alias{filter_eurocordex_filelist}
\title{go through file list and look from which files data should be read}
\usage{
filter_eurocordex_filelist(fList, num_time)
}
\arguments{
\item{fList}{data.frame containing columns filename (char), time_start (int/numeric), and time_stop (int/numeric).}

\item{num_time}{2-element array of numeric or integer}
}
\value{
data.frame containing a filtered version of fList; additional columns are added, which indicate which records have to be read from the netCDF files
}
\description{
go through file list and look from which files data should be read
}
\examples{

# Used by `read_eurocordex_data`. Please look into that function's code for a usage example.


}
