% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eurocordex_correct_time.R
\name{get_eurocordex_correct_time}
\alias{get_eurocordex_correct_time}
\title{correct the input parameter 'time' for read_eurocordex_data()}
\usage{
get_eurocordex_correct_time(time)
}
\arguments{
\item{time}{length-1 or length-2 array of type character, numeric or integer}
}
\value{
numeric or integer array of length 2
}
\description{
correct the input parameter 'time' for read_eurocordex_data()
}
\examples{

# Used by `read_eurocordex_data`. Please look into that function's code for a usage example.


}
