% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eurocordex_filelist.R
\name{get_eurocordex_filelist}
\alias{get_eurocordex_filelist}
\title{Look for files in a specific directory and return a list of files with start and end dates}
\usage{
get_eurocordex_filelist(dir, file)
}
\arguments{
\item{dir}{string: directory name}

\item{file}{string: file name part we look for 'dir/file_YYYYDDMM-YYYYDDMM.nc' files}
}
\value{
data.frame with three columns 'filename' (without path), 'time_start' and 'time_stop' (format: YYYYMMDD)
}
\description{
Look for files in a specific directory and return a list of files with start and end dates
}
\examples{
 
# Used by `read_eurocordex_data`. Please look into that function's code for a usage example.
}
