% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eurocordex_files.R
\name{read_eurocordex_files}
\alias{read_eurocordex_files}
\title{read actual data from netCDF files}
\usage{
read_eurocordex_files(dir, fImport, idxStations, varname)
}
\arguments{
\item{dir}{string: file direcotory}

\item{fImport}{data.frame: containing columns filename, read_idx_start, and read_idx_count}

\item{idxStations}{data.frame: containing columns station_name, idx_lon, and idx_lat}

\item{varname}{variable name to import}
}
\value{
data.frame with first column "time" and last column "time_posix". further columns in between are denoted according to idxStations$station_name; each row contains data of one time step
}
\description{
read actual data from netCDF files
}
\examples{

# Used by `read_eurocordex_data`. Please look into that function's code for a usage example.


}
