% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_shape_station_data.R
\name{read_swat_shape_station_data}
\alias{read_swat_shape_station_data}
\title{Read station data from a dbf file (part of the set of ArcGIS shape files)}
\usage{
read_swat_shape_station_data(file)
}
\arguments{
\item{file}{character: directory and file name}
}
\value{
data.frame: consisting of four columns 'station_name', 'lon', 'lat', and 'elev'
}
\description{
Reads from a *.dbf file
}
\examples{

file <- './my_stations.dbf'

read_swat_shape_station_data(file)
#   station_name   lon   lat elev
# 1          zyx 12.10 53.61   58
# 2          wvu 11.39 53.64   59
# 3          tsr 12.56 53.76   38
# 4          qpo 12.33 54.07   34

}
