% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eurocordex_faketime.R
\name{set_eurocordex_faketime}
\alias{set_eurocordex_faketime}
\title{Replace time data in a data.frame by new time}
\usage{
set_eurocordex_faketime(data, fakeTime, col_time = "time",
  col_time_posix = "time_posix", fakeLength = TRUE)
}
\arguments{
\item{data}{data.frame: contains data and time columns ('time' and 'time_posix'); time_posix column needs to be of type POSIXct}

\item{fakeTime}{POSIXct-array: contains new times to insert}

\item{col_time}{character, optional: set alternative name of column 'time'}

\item{col_time_posix}{character, optional: set alternative name of column 'time_posix'}

\item{fakeLength}{boolean, optional: if data has less entries than the length of fake time => extend data by repeating entries}
}
\value{
data.frame: with 'corrected' time axies
}
\description{
Replace time data in a data.frame by new time
}
\examples{

# see example in README

}
