% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_swat2012_pcp.R
\name{write_input_swat2012_pcp}
\alias{write_input_swat2012_pcp}
\title{Write out precipitation SWAT 2012 meteorology input file}
\usage{
write_input_swat2012_pcp(file, data, stations, station_names,
  col_time_name = "time_posix", col_station_name = "station_name",
  col_lon = "lon", col_lat = "lat", col_elev = "elev")
}
\arguments{
\item{file}{string: directory + file name of output file}

\item{data}{data.frame: precipitation [mm]; contains a column with time stamp and >=1 columns with data; time column needs to be of type POSIXct}

\item{stations}{data.frame : contains station information in four columns: station name, lon coordinate, lat coordinate, and elevation}

\item{station_names}{character-array: contains names of the stations to write out; have to be equal to column names of data}

\item{col_time_name}{character, optional: name of the time stamp column in 'data', default = 'time_posix'}

\item{col_station_name}{character, optional: name of the station names column in 'stations', default = 'station_name'}

\item{col_lon}{character, optional: name of the longitude column in 'stations', default = 'lon'}

\item{col_lat}{character, optional: name of the latitude column in 'stations', default = 'lat'}

\item{col_elev}{character, optional: name of the elevation column in 'stations', default = 'elev'}
}
\description{
Write out precipitation SWAT 2012 meteorology input file
}
\examples{

# see example in README

explanation of the following command:
  - make time column:
      format(data[[time_name]],"\%Y\%m\%d")
  - format the individual values of the data column:
      formatC(as.matrix(my_data[station_names]), format='f', width=8, digits=3, flag='0')
     We need 'as.matrix' because my_data is a data frame. formatC cannot be applied to a 2d data frame.
  - append all formatted data values in one row:
      apply(formatC(...), 1, paste0, collapse = '')
}
